feat: add admin controller and update email templates

+++ src/Controller/competences/ApcController.php
@@ -10,12 +10,15 @@
 namespace App\Controller\competences;
 
 use App\Classes\Apc\ApcStructure;
+use App\Classes\Export\DepartementExport;
 use App\Classes\Import\MyUpload;
 use App\Classes\Import\ReferentielCompetenceImport;
 use App\Controller\BaseController;
 use App\Entity\Constantes;
 use App\Entity\Departement;
 use App\Repository\DepartementRepository;
+use App\Classes\JsonDiffService;
+use App\Utils\Files;
 use Knp\Bundle\SnappyBundle\Snappy\Response\PdfResponse;
 use Knp\Snappy\Pdf;
 use Symfony\Component\HttpFoundation\Request;
@@ -28,12 +31,17 @@ class ApcController extends BaseController
     #[Route("/consulter/{departement}", name:"administration_apc_referentiel_index", methods:["GET"])]
     public function referentiel(ApcStructure $apcStructure, Departement $departement = null): Response
     {
+        if (null === $departement) {
+            throw new \Exception('Departement inconnu');
+        }
+
         $tParcours = $apcStructure->parcoursNiveaux($departement);
         $competences = $departement->getApcCompetences();
         $tComp = [];
         foreach ($competences as $comp) {
             $tComp[$comp->getId()] = $comp;
         }
+
         $competencesParcours = [];
 
         foreach ($tParcours as $key => $parc) {
@@ -55,12 +63,17 @@ class ApcController extends BaseController
     #[Route('/exporter/{departement}', name: 'export_referentiel_competences', methods: ['GET'])]
     public function exportReferentiel(Pdf $knpSnappyPdf, ApcStructure $apcStructure, Departement $departement = null): PdfResponse
     {
+        if (null === $departement) {
+            throw new \Exception('Departement inconnu');
+        }
+
         $tParcours = $apcStructure->parcoursNiveaux($departement);
         $competences = $departement->getApcCompetences();
         $tComp = [];
         foreach ($competences as $comp) {
             $tComp[$comp->getId()] = $comp;
         }
+
         $competencesParcours = [];
 
         foreach ($tParcours as $key => $parc) {
@@ -86,6 +99,105 @@ class ApcController extends BaseController
         );
     }
 
+    #[Route('/exporter-versionning/{departement}', name: 'export_versionning_referentiel_competences', methods: ['GET'])]
+    public function exportVersionReferentiel(
+        Files $files,
+        DepartementExport $departementExport,
+        Pdf $knpSnappyPdf, ApcStructure $apcStructure, Departement $departement = null): Response //PdfResponse
+    {
+        if (null === $departement) {
+            throw new \Exception('Departement inconnu');
+        }
+
+        // version précédente :
+        $fichier = $files->getLastVersionFile($departement);
+        $tabAncien = json_decode(file_get_contents($fichier), true);
+
+        // version courante :
+        $tabActuel = $departementExport->genereJson($departement);
+
+        $diffService = new JsonDiffService();
+        $diffs = $diffService->compare($tabAncien, $tabActuel);
+//        dd($diffs);
+
+        $tParcours = $apcStructure->parcoursNiveaux($departement);
+        $competences = $departement->getApcCompetences();
+        $tComp = [];
+        foreach ($competences as $comp) {
+            $tComp[$comp->getId()] = $comp;
+        }
+
+        $competencesParcours = [];
+
+        foreach ($tParcours as $key => $parc) {
+            $competencesParcours[$key] = [];
+            foreach ($parc as $k => $v) {
+                $competencesParcours[$key][] = $tComp[$k];
+            }
+        }
+
+        $html =  $this->renderView('competences/export-versionning-referentiel.html.twig',[
+            'competencesParcours' => $competencesParcours,
+            'departement' => $departement,
+            'competences' => $competences,
+            'parcours' => $departement->getApcParcours(),
+            'parcoursNiveaux' => $tParcours,
+            'diffs' => $diffs,
+        ]);
+
+        return new PdfResponse(
+            $knpSnappyPdf->getOutputFromHtml($html, [
+                'orientation'=>'Landscape'
+            ]),
+            'referentiel-competence-'.$departement->getSigle().'.pdf'
+        );
+    }
+
+    #[Route('/voir-versionning/{departement}', name: 'voir_versionning_referentiel_competences', methods: ['GET'])]
+    public function voirVersionReferentiel(
+        Files $files,
+        DepartementExport $departementExport,
+        ApcStructure $apcStructure, Departement $departement = null): Response //PdfResponse
+    {
+        if (null === $departement) {
+            throw new \Exception('Departement inconnu');
+        }
+
+        // version précédente :
+        $fichier = $files->getLastVersionFile($departement);
+        $tabAncien = json_decode(file_get_contents($fichier), true);
+
+        // version courante :
+        $tabActuel = $departementExport->genereJson($departement);
+
+        $diffService = new JsonDiffService();
+        $diffs = $diffService->compare($tabAncien, $tabActuel);
+        $tParcours = $apcStructure->parcoursNiveaux($departement);
+        $competences = $departement->getApcCompetences();
+        $tComp = [];
+        foreach ($competences as $comp) {
+            $tComp[$comp->getId()] = $comp;
+        }
+
+        $competencesParcours = [];
+
+        foreach ($tParcours as $key => $parc) {
+            $competencesParcours[$key] = [];
+            foreach ($parc as $k => $v) {
+                $competencesParcours[$key][] = $tComp[$k];
+            }
+        }
+
+        return  $this->render('competences/voir-versionning-referentiel.html.twig',[
+            'competencesParcours' => $competencesParcours,
+            'departement' => $departement,
+            'competences' => $competences,
+            'parcours' => $departement->getApcParcours(),
+            'parcoursNiveaux' => $tParcours,
+            'diffs' => $diffs,
+        ]);
+    }
+
     #[Route("/import", name:"administration_apc_referentiel_import", methods:["GET","POST"])]
     public function import(
         DepartementRepository $departementRepository,
@@ -93,14 +205,11 @@ class ApcController extends BaseController
         ReferentielCompetenceImport $diplomeImport,
         Request $request
     ): Response {
-        if ($request->isMethod('POST')) {
-            if (null !== $this->getDepartement()) {
-                $fichier = $myUpload->upload($request->files->get('fichier'), 'temp/', ['xml', 'xlsx']);
-                $diplomeImport->import($this->getDepartement(), $fichier, $request->request->get('typeFichier'));
-                unlink($fichier);
-                $this->addFlashBag(Constantes::FLASHBAG_SUCCESS, 'Maquette importée avec succès');
-            }
-
+        if ($request->isMethod('POST') && null !== $this->getDepartement()) {
+            $fichier = $myUpload->upload($request->files->get('fichier'), 'temp/', ['xml', 'xlsx']);
+            $diplomeImport->import($this->getDepartement(), $fichier, $request->request->get('typeFichier'));
+            unlink($fichier);
+            $this->addFlashBag(Constantes::FLASHBAG_SUCCESS, 'Maquette importée avec succès');
         }
 
         return $this->render('import_referentiel/index.html.twig', [
